+++ src/main/java/com/umc/EveryWear/domain/product/service/command/ProductCommandServiceImpl.java
@@ -25,7 +25,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.BodyInserters;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
-import reactor.core.publisher.Mono;
 import org.springframework.dao.DataIntegrityViolationException;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
@@ -34,10 +33,6 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
 
 @Slf4j
 @Service
@@ -58,8 +53,6 @@ public class ProductCommandServiceImpl implements ProductCommandService {
     private String fastApiBaseUrl;
 
     private static final int CRAWL_TIMEOUT_SECONDS = 120;
-    private static final int REDIRECT_RESOLVE_TIMEOUT_SECONDS = 10;
-    private static final int PRODUCT_NUM_TOTAL_LENGTH = 15;
 
     private ShoppingMall detectShoppingMall(String url) {
         if (url == null) return null;
@@ -95,10 +88,10 @@ public class ProductCommandServiceImpl implements ProductCommandService {
             }
 
             // 1. 리다이렉트가 필요하면 최종 URL 확보
-            String finalUrl = resolveRedirect(productUrl);
+            String finalUrl = ProductUrlUtil.resolveRedirect(productUrl);
 
             // 2. 최종 URL에서 상품 고유 번호(product_num) 추출
-            Long productNum = extractProductNumFromFinalUrl(finalUrl, mall);
+            Long productNum = ProductUrlUtil.extractProductNumFromFinalUrl(finalUrl, mall);
 
             // 3. product_num으로 DB 존재 여부 판단
             if (productNum != null) {
@@ -111,7 +104,7 @@ public class ProductCommandServiceImpl implements ProductCommandService {
 
             // 4. DB에 없으면 크롤링 후 상품 저장 및 user_product 연결
             ProductCrawlingData crawlerData = crawlProduct(productUrl, mall);
-            String canonicalUrl = canonicalizeProductUrl(crawlerData.getProductUrl(), mall);
+            String canonicalUrl = ProductUrlUtil.canonicalizeProductUrl(crawlerData.getProductUrl(), mall);
             crawlerData.setProductUrl(canonicalUrl);
             try {
                 return createProductAndLink(user, crawlerData);
@@ -153,80 +146,6 @@ public class ProductCommandServiceImpl implements ProductCommandService {
                 .orElseThrow(() -> new ProductException(ProductErrorCode.CRAWLING_FAILED));
         productRepository.updateUpdatedAt(existing.getProductId());
         return resolveOrLinkUserProduct(userId, user, existing, true);
-    }
-
-    // 요청 URL을 리다이렉트한 최종 URL로 변환
-    private String resolveRedirect(String url) {
-        try {
-            HttpClient client = HttpClient.newBuilder()
-                    .followRedirects(HttpClient.Redirect.ALWAYS)
-                    .connectTimeout(Duration.ofSeconds(REDIRECT_RESOLVE_TIMEOUT_SECONDS))
-                    .build();
-
-            HttpRequest request = HttpRequest.newBuilder()
-                    .uri(URI.create(url))
-                    .GET()
-                    .build();
-
-            HttpResponse<Void> response = client.send(request, HttpResponse.BodyHandlers.discarding());
-            URI finalUri = response.uri();
-            return finalUri != null ? finalUri.toString() : url;
-        } catch (Exception e) {
-            log.warn("리다이렉트 확인 실패, 원본 URL 사용: {} - {}", url, e.getMessage());
-            return url;
-        }
-    }
-
-    // 최종 URL에서 상품 ID를 파싱한 뒤, DB 저장 형식으로 변환
-    private Long extractProductNumFromFinalUrl(String finalUrl, ShoppingMall mall) {
-        if (finalUrl == null || mall == null) return null;
-        try {
-            String path = new URI(finalUrl).getPath();
-            if (path == null) return null;
-
-            Long rawId = null;
-            int prefixDigit;
-            switch (mall) {
-                case MUSINSA:
-                    rawId = extractProductId(path, "/products/");
-                    prefixDigit = 1;
-                    break;
-                case ZIGZAG:
-                    rawId = extractProductId(path, "/catalog/products/");
-                    if (rawId == null) rawId = extractProductId(path, "/p/");
-                    prefixDigit = 2;
-                    break;
-                case WCONCEPT:
-                    rawId = extractProductId(path, "/Product/");
-                    if (rawId == null) rawId = extractProductId(path, "/product/");
-                    prefixDigit = 4;
-                    break;
-                case CM29:
-                    rawId = extractProductId(path, "/products/");
-                    prefixDigit = 3;
-                    break;
-                default:
-                    return null;
-            }
-            if (rawId == null) return null;
-            return formatProductNum(rawId, prefixDigit);
-        } catch (Exception e) {
-            log.warn("상품 번호 추출 실패 (mall: {}, url: {}): {}", mall, finalUrl, e.getMessage());
-            return null;
-        }
-    }
-
-    // 크롤러와 동일한 15자리 product_num 형식으로 포맷
-    private static Long formatProductNum(long productId, int prefixDigit) {
-        String idStr = String.valueOf(productId);
-        int zerosNeeded = PRODUCT_NUM_TOTAL_LENGTH - 1 - idStr.length();
-        if (zerosNeeded < 0) return productId;
-        String formatted = prefixDigit + "0".repeat(zerosNeeded) + idStr;
-        try {
-            return Long.parseLong(formatted);
-        } catch (NumberFormatException e) {
-            return null;
-        }
     }
 
     // 기존 Product에 대한 UserProduct 연결 또는 updatedAt 갱신 후 ImportDTO 반환 (PRODUCT202용)
@@ -362,84 +281,6 @@ public class ProductCommandServiceImpl implements ProductCommandService {
             try { return Long.parseLong(n.asText()); } catch (NumberFormatException e) { return null; }
         }
         return null;
-    }
-
-    // 정규화 URL 생성
-    private String canonicalizeProductUrl(String url, ShoppingMall mall) {
-        if (url == null || mall == null) return url;
-
-        try {
-            URI uri = new URI(url);
-            String path = uri.getPath(); // 쿼리스트링이 제거된 순수 path
-            if (path == null) return url;
-
-            switch (mall) {
-                case MUSINSA: {
-                    Long id = extractProductId(path, "/products/");
-                    if (id != null) {
-                        return "https://www.musinsa.com/products/" + id;
-                    }
-                    break;
-                }
-                case ZIGZAG: {
-                    // Case1: https://zigzag.kr/catalog/products/{id}
-                    Long id = extractProductId(path, "/catalog/products/");
-                    if (id != null) {
-                        return "https://zigzag.kr/catalog/products/" + id;
-                    }
-                    // Case2: https://zigzag.kr/p/{id}
-                    id = extractProductId(path, "/p/");
-                    if (id != null) {
-                        return "https://zigzag.kr/p/" + id;
-                    }
-                    break;
-                }
-                case WCONCEPT: {
-                    // m.wconcept / www.wconcept 모두 /Product/{id} 형태로 정규화
-                    Long id = extractProductId(path, "/Product/");
-                    if (id == null) {
-                        id = extractProductId(path, "/product/");
-                    }
-                    if (id != null) {
-                        return "https://www.wconcept.co.kr/Product/" + id;
-                    }
-                    break;
-                }
-                case CM29: {
-                    Long id = extractProductId(path, "/products/");
-                    if (id != null) {
-                        return "https://www.29cm.co.kr/products/" + id;
-                    }
-                    break;
-                }
-            }
-        } catch (Exception e) {
-            log.warn("상품 URL 정규화 실패 (mall: {}, url: {}): {}",
-                    mall != null ? mall.name() : "null", url, e.getMessage());
-        }
-
-        // 패턴에 맞지 않으면 원본 URL 유지
-        return url;
-    }
-
-    // path 문자열에서 prefix 뒤에 이어지는 숫자 부분을 상품 ID로 파싱한다.
-    // 예) path = "/products/3073492?xxx", prefix="/products/" -> 3073492
-    private static Long extractProductId(String path, String prefix) {
-        int idx = path.indexOf(prefix);
-        if (idx < 0) return null;
-
-        int start = idx + prefix.length();
-        int end = start;
-        while (end < path.length() && Character.isDigit(path.charAt(end))) {
-            end++;
-        }
-        if (end == start) return null; // 숫자가 하나도 없으면 실패
-
-        try {
-            return Long.parseLong(path.substring(start, end));
-        } catch (NumberFormatException e) {
-            return null;
-        }
     }
 
     @Override
